create patch helper and move functions to run.py
